+++ src/scale/scaletype.rs
@@ -574,7 +574,7 @@ pub(super) fn advance(
         Simplification::Exact => interval.transpose_pitch_with_options(pitch, false, Some(4)),
         Simplification::MostCommon => {
             let mut transposed = interval.transpose_pitch_with_options(pitch, false, Some(4))?;
-            if transposed.accidental_or_natural().alter() != 0.0 {
+            if transposed.accidental().alter() != 0.0 {
                 transposed.simplify_enharmonic_in_place(true)?;
             }
             Ok(transposed)
@@ -587,7 +587,7 @@ pub(super) fn advance(
 fn max_alter(pitches: &[Pitch]) -> crate::defaults::IntegerType {
     pitches
         .iter()
-        .map(|pitch| pitch.accidental_or_natural().alter().abs() as crate::defaults::IntegerType)
+        .map(|pitch| pitch.accidental().alter().abs() as crate::defaults::IntegerType)
         .max()
         .unwrap_or(0)
 }